Adds a helper to prepare the response
One of the HN endpoints, `/v0/maxitem.json`, returns a single string of digits as the response body. `Faraday::Response::ParseJson` uses `JSON.parse` to parse the response body and subsequently raises a parsing error when encountering an 'invalid token'. It looks like a single value is considered perfectly valid JSON now (http://tools.ietf.org/html/rfc7159#section-2), but it throws a wrench in just using the stock middleware to prepare the response.

I ended up using a custom `prepare` helper in the `Request` module to handle this. As I think the way I'm handling this is a little too hackish, I'd like to look for a better way to deal with this at a later date.